+++ src/main/java/com/playus/twpservice/domain/party/dto/partycreate/PartyCreateRequest.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.party.dto.party_create;
+package com.playus.twpservice.domain.party.dto.partycreate;
 
 import com.playus.twpservice.global.validation.ValidEnum;
 import com.playus.twpservice.global.validation.ValidEnumList;
